securisation du restTemplate entre agent et client

+++ agent_back/src/main/java/com/ensa/e_banking/service/ClientService.java
@@ -8,16 +8,25 @@ import java.util.Date;
 import java.util.List;
 
 import javax.mail.MessagingException;
+import javax.persistence.Entity;
+import javax.servlet.http.HttpServletRequest;
 
 import com.ensa.e_banking.dao.CompteRepository;
 import com.ensa.e_banking.dao.OperationRepository;
 
 import com.ensa.e_banking.entities.Compte;
+import com.ensa.e_banking.entities.Operation;
+import com.ensa.e_banking.security.SecurityConstants;
+import com.sun.org.apache.xpath.internal.operations.Bool;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.annotation.Bean;
 import org.springframework.core.ParameterizedTypeReference;
+import org.springframework.http.HttpEntity;
+import org.springframework.http.HttpHeaders;
 import org.springframework.http.HttpMethod;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
 import org.springframework.transaction.annotation.Transactional;
 import org.springframework.web.bind.annotation.CrossOrigin;
@@ -37,14 +46,15 @@ import org.springframework.web.client.RestTemplate;
 @CrossOrigin("*")
 public class ClientService {
 
-	@Autowired
-	private HomeController homeController;
+
 
     @Autowired
     RestTemplate restTemplate;
 
     private String url = "http://localhost:8082";
 	private String urla="http://localhost:8083";
+
+	HttpHeaders headers=new HttpHeaders();
 
 //	@Autowired
 //	private ClientMetier clientMetier;
@@ -58,11 +68,19 @@ public class ClientService {
 	@Autowired
 	private OperationRepository operationRepository;
 
+	@Autowired
+	private  HomeController homeController;
+
 
 //
 	@RequestMapping(value="/client/ajoutClient",method=RequestMethod.POST)
-	public Client saveClient(@RequestBody Client client) {
-		return restTemplate.postForObject(url+"/client/ajoutClient", client, Client.class);
+	public Client saveClient(@RequestBody Client client,HttpServletRequest request) {
+
+
+		headers.set(SecurityConstants.HEADER_STRING,
+				SecurityConstants.TOKEN_PREFIX+request.getHeader(SecurityConstants.HEADER_STRING));
+		HttpEntity<Client> req = new HttpEntity<>(client,headers);
+		return restTemplate.postForObject(url+"/client/ajoutClient", req, Client.class);
 	}
 //
 //	/*@RequestMapping(value="/client/listClient",method=RequestMethod.GET)
@@ -91,19 +109,24 @@ BCryptPasswordEncoder bCryptPasswordEncoder=new  BCryptPasswordEncoder();
 //
 //
 	@RequestMapping(value="/client/update/{id}",method=RequestMethod.PUT)
-	public Boolean updateClient(@PathVariable long id,@RequestBody Client client){
-		restTemplate.put(url+"client/update/"+id, client);
+	public Boolean updateClient(@PathVariable long id,@RequestBody Client client,HttpServletRequest request){
+
+		headers.set(SecurityConstants.HEADER_STRING,
+				SecurityConstants.TOKEN_PREFIX+request.getHeader(SecurityConstants.HEADER_STRING));
+		HttpEntity<Client> entity = new HttpEntity<Client>(client,headers);
+
+		restTemplate.put(url+"client/update/"+id, entity);
 		return true;
 	}
 //
 	@RequestMapping(value="/client/{id}",method=RequestMethod.GET)
-	public Client getClientById(@PathVariable Long id) throws ParseException {
+	public Client getClientById(@PathVariable Long id,HttpServletRequest request) throws ParseException {
 
 
 
 
 //		return restTemplate.getForObject(url+"/client/"+id, Client.class);
-		List<Client> list = getClients();
+		List<Client> list = getClients(request);
 		for(Client cl : list){
 
 			if(cl.getId() == id) {
@@ -116,14 +139,28 @@ BCryptPasswordEncoder bCryptPasswordEncoder=new  BCryptPasswordEncoder();
 
 	@Transactional
 	@RequestMapping(value="/client/delete/{id}",method=RequestMethod.DELETE)
-	public boolean supprimer(@PathVariable long id){
+	public boolean supprimer(@PathVariable long id,HttpServletRequest request){
 
 		Compte compte=compteRepository.findCompteByIdClient(id);
+
+     if(operationRepository.findOperationByIdCompte(compte.getRib()) !=null){
 		operationRepository.deleteOperationByCompte(compte.getRib());
-		compteRepository.deleteCompteByIdClient(id);
-	    restTemplate.delete(url+"client/delete/"+id);
+		 compteRepository.deleteCompteByIdClient(id);
+     System.out.println("opera not null");}
+     else{  compteRepository.deleteCompteByIdClient(id);}
+
+
+
+		headers.set(SecurityConstants.HEADER_STRING,
+				SecurityConstants.TOKEN_PREFIX+request.getHeader(SecurityConstants.HEADER_STRING));
+		HttpEntity<Client> entity = new HttpEntity<Client>(headers);
+
+
+	  // restTemplate.delete(url+"client/delete/"+id,entity);
+		restTemplate.exchange(url+"client/delete/"+id, HttpMethod.DELETE,entity, Boolean.class);
 		String act= "l'agent ID : "+ homeController.currentAgent().getId() + " a supprimé le client ID : "+ id +
 				" ainsi que son compte RIB : " + compte.getRib();
+
 		restTemplate.postForObject(urla + "/supprimerClient", act, String.class);
 	    return true;
 	}
@@ -138,11 +175,13 @@ BCryptPasswordEncoder bCryptPasswordEncoder=new  BCryptPasswordEncoder();
 //	/*
 //
 	@RequestMapping(value="/agent/chercher/{mc}",method=RequestMethod.GET)
-	public List<Client> chercher(@PathVariable String mc){
+	public List<Client> chercher(@PathVariable String mc,HttpServletRequest request){
 		System.out.println(mc);
-
+		headers.set(SecurityConstants.HEADER_STRING,
+				SecurityConstants.TOKEN_PREFIX+request.getHeader(SecurityConstants.HEADER_STRING));
+		HttpEntity<Client> entity = new HttpEntity<Client>(headers);
 			ResponseEntity<List<Client>> response = restTemplate.exchange(
-					url+"/client/recherche/"+mc, HttpMethod.GET, null, new ParameterizedTypeReference<List<Client>>() {}
+					url+"/client/recherche/"+mc, HttpMethod.GET, entity, new ParameterizedTypeReference<List<Client>>() {}
 			);
 
 			List<Client> list = response.getBody();
@@ -151,12 +190,24 @@ BCryptPasswordEncoder bCryptPasswordEncoder=new  BCryptPasswordEncoder();
 
 
     @RequestMapping(value="/client/list",method = RequestMethod.GET)
-    public List<Client> getClients(){
+    public List<Client> getClients(HttpServletRequest request){
+
+
+		headers.set(SecurityConstants.HEADER_STRING,
+				SecurityConstants.TOKEN_PREFIX+request.getHeader(SecurityConstants.HEADER_STRING));
+		HttpEntity<Client> entity = new HttpEntity<Client>(headers);
+		ResponseEntity<List<Client>> response = restTemplate.exchange(
+				url+"/client/list", HttpMethod.GET, entity, new ParameterizedTypeReference<List<Client>>() {}
+		);
+		List<Client> list = response.getBody();
+		return list;
+
+		/*
         ResponseEntity<List<Client>> response = restTemplate.exchange(
                 url+"/client/list", HttpMethod.GET, null, new ParameterizedTypeReference<List<Client>>() {}
         );
         List<Client> list = response.getBody();
-        return  list;
+        return  list;*/
     }
 			
 }
